Added a clamp function into `math` library. Changed the way screens data is structured.
Changed the way menu is updated so you can update individual items of it.
Defining labels for items is optional if you have defined an `update` function for it in which you update the label(makes the code shorter and easier to maintain i believe).
With this change I was able to make `Screen` option in `Window Options` as a separate item, upon changed it updates the `Resolution` item and it adjust proportionally to selected monitor if it differs with resolution to a previous one.

Did a slight refactoring(hope it's not too bad for you >w>;;).